fix: harden image decoders, OPDS search URLs, and latent overflow paths
- OpdsBookBrowserActivity: strip optional OpenSearch placeholders
  ({name?}) after searchTerms substitution and tidy dangling
  separators, per the OpenSearch spec
- validateImageDimensions: static, and rejects dims over INT16_MAX so
  the int16_t casts in getDimensionsStatic cannot wrap; JPEG/PNG
  static dimension queries now validate before truncating
- PngToFramebufferConverter: 64-bit row-buffer sizing with a pitch cap
- PixelCache: widen bytesPerRow*height before write
- Section: LUT bounds checks computed in uint64 (no uint32 wrap)
- BleWifiProvisioner: log + bail on null state mutex

+++ lib/Epub/Epub/converters/PixelCache.h
@@ -67,10 +67,11 @@ struct PixelCache {
     uint16_t h = height;
     cacheFile.write(&w, 2);
     cacheFile.write(&h, 2);
-    cacheFile.write(buffer, bytesPerRow * height);
+    cacheFile.write(buffer, static_cast<size_t>(bytesPerRow) * static_cast<size_t>(height));
     cacheFile.close();
 
-    LOG_DBG("IMG", "Cache written: %s (%dx%d, %d bytes)", cachePath.c_str(), width, height, 4 + bytesPerRow * height);
+    LOG_DBG("IMG", "Cache written: %s (%dx%d, %zu bytes)", cachePath.c_str(), width, height,
+            4 + static_cast<size_t>(bytesPerRow) * static_cast<size_t>(height));
     return true;
   }
